Remove several fields from `DataSerializer`
`original_chunk_type` and `compressed_chunk_type` have no effect, because
the task creation code overwrites these values based on `use_zip_chunks`.
With `size`, the code relies on the default of zero, and attempting to
overriding it will just lead to an incorrect value being saved to the
database.