Split off documentation building into its own pipeline
It has occurred to me that the documentation build happens inside the main
workflow, which is configured to ignore changes to the docs when determining
whether to run. So if a PR only changes the docs... we don't build the docs.
That makes no sense.

Fix this by extracting the documentation job into its own workflow, that
always runs. The downside of this is that we have to duplicate the work of
generating the SDK, but IMO, it's an acceptable sacrifice.

A side effect of this is that a failure to build the docs no longer prevents
the publishing of Docker images, but I don't think that was needed in the
first place.