Streamline API endpoint summaries (#7685)
The current summaries are much too verbose. This makes it hard to find
the endpoint you're looking for in the API docs, because it takes longer
to read & understand each summary.

Rewrite them to be more brief. Move any details that are too long to
keep, but too important to omit, to the description.