Docs: fix broken internal links and ensure they stay fixed (#7695)
We have quite a few broken links, which is unsurprising, as Hugo does
not help detect them.

This patch adds a custom shortcode (`ilink`) which validates the target
path and then renders as a link. It rewrites most internal links to use
this shortcode, and fixes those which were broken.

Originally, I was just going to use the builtin `relref` shortcode to
build the URLs, but it turns out that that does not play well with
remark. It becomes confused and thinks you forgot to specify the URL
altogether. `ilink` replaces the entire link, so remark has nothing to
complain about. `ilink` supports a default link text, although I haven't
yet used this anywhere.

I also "fortified" the "about" link in `navbar.html`, although templates
don't support shortcodes, so I just used the `relref` function.

A few links I had to keep as-is for various reasons:

* Links from the `docs/_index.md`, because they are inside a shortcode,
  and as far as I can tell, you can't put shortcodes in shortcode
  parameters.

* The refresh URLs in `home.html` and `404.html`, for complicated
  reasons.

* Links to autogenerated API documentation pages due to some known
  problems with generating those pages for previous releases.

Finally, I had to update `build_docs.py`, because it neglected to check
out `page_404.html` when rendering the docs for previous versions, and
the new version of `page_404.html` will not render without the new
shortcode.