Fix spurious failures in `case_17_lock_hide_features.js`
69682f25 changed the sidebar layout in such a way that mousing over the
"Switch hidden property" button for an object can show a tooltip that
obscures the "Switch hidden property for all" button. This is exactly what
happens in this test.

The test is supposed to work around this, by hiding the tooltips before each
section, but the workaround is insufficiently robust, because it can fire
before the tooltip is shown.

Fix this by taking another workaround from
`issue_2485_navigation_empty_frames.js`, moving it into the `hideTooltips`
command, and then using that command in `case_17_lock_hide_features.js`.