Split cvat/apps/iam/permissions.py into per-app permissions modules (#7734)
This improves the dependency graph of apps with regards to IAM and other
apps. IAM is a generic app, so it should not depend on specific apps,
like `webhooks`.

It's also easier to work with a bunch of smaller files rather than one
giant one.

Having permissions classes in separate modules introduces the risk that
they will never be loaded directly. To mitigate that, add a function to
ensure permissions are loaded and call it from every app.

In a future PR, I plan to move the `.rego` files into their respective
apps, as well.