Modernize Rego syntax
Open Policy Agent v0.59 introduced a new directive (`import rego.v1`) that
ensures that the file is compatible with OPA v1 (to be released in the
future).

Add this directive to all Rego files and update the syntax accordingly.
Which involves the following:

* Rewrite all rules to use the `if` keyword, which is now mandatory.

* Where appropriate, use the `in` keyword, which is now available without a
  future import. It's not mandatory, but it looks much nicer.

In addition, update Regal to the latest version, which now enforces the use
of `import rego.v1` by default.