Use CPU PyTorch for testing the SDK
Not only is the GPU version of PyTorch much bigger than the CPU version,
but it also pulls in CUDA, which is enormous. We don't (and can't) use any
GPU features in our tests, so we don't need the GPU version.

Eliminating it saves ~4GB of disk space, which is precious when it comes to
the GitHub actions runners.