helm-chart: prevent Traefik from ignoring the backend ingress rule
There is a condition that may occur during Kubernetes deployment, where the
frontend service already has an endpoint (i.e. the frontend pod), but the
backend service does not. For example, the backend pod may not have started
yet or the service controller may not have had time to react to the backend
pod.

In this case, when Traefik serves a request with an `/api/...` path, it will
see that it matches the `/api` rule, but since the corresponding service has
no endpoints, it will _skip_ that rule and try other rules. And since the
`/` rule matches everything, it will then route the request to the frontend.

This is confusing and unhelpful, and more importantly, it makes health
checks return the wrong result. Since the frontend will serve `index.html`
to every request, a request to `/api/server/health/` or `/api/server/about`
will return a 200 code, even though the server isn't actually up.

Because of this bug, I have observed weird failures in the Helm workflow,
where the "Wait for CVAT to be ready" step completes, but CVAT is not
actually ready. (FYI: The failures I've seen are actually in a private repo,
but the failure condition could occur in this repo too. It's just more
likely in a private repo, because GitHub uses smaller runners in private
repos.)

The fix is simple: use the `allowEmptyServices` Traefik setting, which
disables the rule skipping behavior. With this setting on, Traefik will
return a 503 response for backend URLs until the backend service gains an
endpoint.