Fix a non-deterministic webhook test
`test_two_project_webhooks_intersection` is supposed to trigger each webhook
once. However, the first one of these webhooks actually gets triggered
twice, because creating a task causes the project's `updated_date` to be
bumped, which triggers an `update:project` event.

The test still passes a lot of the time (I guess because the second delivery
doesn't appear immediately?), but sometimes it fails. It's very easy to make
it fail consistently, though - just add a `sleep(5)` before the
`get_deliveries` calls.

Fix this by changing the first webhook's second event to something that will
not be triggered.