Remove `ModelKind.CLASSIFIER`
I'd like the "kind" field in the API to identify the function's "signature",
or the types of values it receives as input and produces as output.
Classifiers have the same signature as detectors, so `classifier` is a
redundant value.

Besides improving semantic purity, removing this redundant value simplifies
the UI code.

The only meaningful difference between how the UI handles classifiers, as
compared to detectors, is that it shows the word "classifier" in the model
modal, which can be helpful. But we can achieve the same thing by
examining the function's `return_type` field. This lets us give a special
label to segmentation functions, as well.

"classifier" can't actually be returned by `/api/lambda/functions`, but it
_can_ be returned by the RoboFlow/Hugging Face function API in CVAT
Enterprise. So we'll need a small compatibility shim to transform this value
to "detector" until I fix that API to stop returning it too.