Layer tus-js-client on top of Axios
This ensures that requests coming from tus-js-client have the same defaults
as the ones coming from the rest of the UI. In particular, this ensures that
TUS requests include the `X-CSRFTOKEN` header.

Currently, this doesn't matter much, because TUS requests are authenticated
using the token. However, I'd like to get rid of token authentication in the
UI, after which `X-CSRFTOKEN` will become important.