Prolong user sessions while they are used
Currently, this shouldn't have any visible effect, because the UI uses
token authentication alongside session cookies, and the tokens last
indefinitely. However, I'd like to end this practice and rely solely
on session cookies.

When that's implemented, the user will get logged out as soon as the session
cookie expires, or the server-side session data expires (which should happen
at the same time). This will irritate users if it happens too often (or
worse, in the middle of their work). Therefore, we should prolong a session
as long as it is used.