Prepare auto-annotation RQ jobs for migration to the general request API
* Define an RQ ID format for auto-annotation jobs, and make use of it.

* Add a `function_id` field to `RequestSerializer`, so that the general
  request API can expose the same information as the lambda request API.
  (In truth, the lambda request API also exposes the "threshold" field,
  but the UI doesn't use it, and I don't see the point in having it.)

Note that this doesn't actually _enable_ the general request API for
auto-annotation requests. This is because a similar patch needs to first be
applied to the Enterprise version, otherwise requests for Roboflow/Hugging
Face jobs will be invisible.

This also fixes a bug where a failed auto-annotation request will prevent
a new one from being started.