Switch to v4 of artifact actions
v3 is due to stop working on 2024-11-30:
<https://github.blog/changelog/2024-04-16-deprecation-notice-v3-of-the-artifact-actions/>.

Note that artifact-upload@v4 no longer allows multiple jobs of the same
workflow to upload an artifact with the same name, thus I had to rename
several of the artifacts. In the case of coverage results, re-merge them at
the end to avoid clutter on the workflow summary page. I didn't bother with
merging other artifacts, because they're only produced in case of failure,
so I'm not expecting them to create much clutter.

Delete the schema upload step entirely, because I don't think it works (it's
inside the wrong job), and it doesn't seem that useful anyway.