Harden access control in several PATCH endpoints (#8493)
Currently several permission classes dynamically determine scopes for
`PATCH` endpoints, depending on request body fields. This is error-prone
for a couple of reasons:

* If you forget to add a default scope like `Scopes.VIEW`, then anyone
can view the resource by submitting a `PATCH` request with an empty
body.

* If you add a field to a resource, but forget to define which scope it
should be associated with, then anyone can modify this field on any
instance of this resource.

Both of these have happened previously; see
<https://github.com/cvat-ai/cvat/security/advisories/GHSA-gxhm-hg65-5gh2>.

Attempt to harden the scope determination logic for such endpoints, so
that such errors could not occur. Specifically:

* Factor out the common logic (including the default scope) into a
function.

* Instead of checking for each known field in the request body, check
every field in the body against a map of known fields, and abort if the
field does not map to any scopes.

A side effect of this change is that request bodies that set certain
fields (such as `owner_id`) to the values they already have did not
previously require the appropriate `UPDATE_*` scope, but now they do.
This is for simplicity of implementation. I see no need to preserve the
current behavior, since a user who does not have the requisite
permission can simply omit the field from the request.

Another side effect is that request bodies with unknown fields are now
rejected, whereas previously such fields would be ignored. I think
that's probably a good thing, since we can't properly fulfill such a
request anyway - can't set a field that doesn't exist.

This change detected a broken test in the unit test suite, so fix that.