Fixed quality reports and immediate feedback on jobs with start frame and step (#8551)
To address the identified issue, we likely need to store relative frame
numbers in self._required_frames.

Explanation:
The only place where self._required_frames is used is in the method:

```
def _is_frame_required(self, frame):
    return self._required_frames is None or frame in self._required_frames
```

This method is called in:

```
def get_included_frames(self):
    return set(
        i for i in self.rel_range
        if not self._is_frame_deleted(i)
        and not self._is_frame_excluded(i)
        and self._is_frame_required(i)
    )

```
Currently, the implementation tries to check whether a relative frame
number exists in a set that contains absolute frame numbers, which
doesn't work if start_frame or frame_step is defined when the task is
created. This causes failures in both quality reports and immediate
feedback, as well as possibly other functionalities that rely on
get_included_frames.


Another issue with the current code is a logical inconsistency:

```
abs_range = self.abs_range
self._required_frames = set(
     self.abs_frame_id(frame) for frame in self._required_frames
     if frame in abs_range
)
```

Here, abs_frame_id expects a relative frame number and converts it to an
absolute frame number, but we first check if the relative frame number
from self._required_frames is already in the absolute range, which is
incorrect.