Add a mechanism for statically defining periodic background jobs (#8552)
My main motivation was really to add a mechanism to statically define
periodic jobs in the project configuration (which I need to do for a
feature I'm working on). Such a mechanism is needed for a few reasons:

1. There's no good place to put code that creates periodic jobs. You
   could do it on app initialization, but that means you have to run Redis
   to use `manage.py`, which is inconvenient for development; and it causes
   much more Redis traffic than necessary, since the code will run whenever
   any backend service starts (or restarts).
   This patch fixes this by adding a new management command,
   `syncperiodicjobs`, that runs once per deployment.

2. It allows to find and delete jobs that are no longer needed. This is
   done by recording all created jobs in a Redis set. Without this, even
   when you delete the code that creates a job, the job itself would
   persist in Redis forever and require manual intervention to delete.

3. It contains common logic to recreate a job when its parameters
   change.

Besides the general mechanism, I also added one concrete job that cleans
up expired Django sessions. This helps to demonstrate the mechanism, and
is useful in its own right.