Allow users to start auto-annotation again if the previous request fails
Currently, if an auto-annotation request fails, the UI doesn't re-enable
the "Automatic annotation" menu item. This seems like a mistake, since
at that point there's no reason why a new request could not be started -
in fact, it's possible to do if you just refresh the page.

Fix this by only disabling the menu item if the current AR is still in
progress.