Update the Docker image handling code in `schedule.yml`
It currently uses an auxiliary Docker Hub account to temporarily store the
images, but all other workflows have long ago switched to using workflow
artifacts. Switch this workflow too, by replacing all image-handling code
with the equivalents from `main.yml`.

The main reason to do this is to fix a bug in the workflow: the
`unit_testing` job tags the `cvat/*` images as `:latest`, but the Compose
file wants the `:dev` tag. So the images end up being redownloaded from
Docker Hub.

There are also a few other reasons why the new version is better:

* It avoids unnecessary complexity, viz. the usage of Docker Hub, an
  external service.

* It's consistent with the other pipelines, which should make future changes
  easier.

* It's much less code.