Make the `--insecure` CLI option test less hacky
Test it by starting an HTTPS server (that forwards all requests to CVAT) and
attempting to access that server via the CLI. This allows us to test that
the option actually works end-to-end, and avoids dependencies on CLI
internals.

The new test requires `main` to catch the exception that is raised in the
case of a certificate error, so implement that.