Fix return type annotations for functions used with @contextmanager
Some of them are annotated with an `Iterator` return type. However...

It just occurred to me that `@contextmanager` cannot work with a function
that returns a plain iterator, since it relies on the generator class's
`throw` method. `contextmanager` is defined in typeshed as accepting an
iterator-returning function, but that appears to be a bug:
<https://github.com/python/typeshed/issues/2772>.

Change all such annotations to a `Generator` type instead.

Some annotations are also broken in other ways; fix them too.