Add a quality setting for avoiding using bbox in point group matching (#8634)
In some cases point groups do not represent a single object or a
bounding box attached to the group does not represent the object. A new
quality check option is added for such cases. The new option allows to
control whether to use the bbox for point group comparisons or not.

- Added a new quality setting to enable or disable using of group bbox
  for point group comparisons
- Improved OKS sigma description