SDK/CLI: Add ability to call auto-annotation functions with a custom threshold
My main motivation is to support a future feature, but I think this is a
good thing in its own right.

While it's already possible to create an AA function that lets you customize the
threshold (by adding a creation parameter), confidence scoring is very
common in detection models, so it makes sense to make this easier to
support, both for the implementer of the function, and for its user.