Make SDK models pickleable
Currently, they aren't (or rather, they can be pickled, but unpickling
fails). This is due to a small quirk of how the model classes work, and is
easily worked around.

In addition, don't create a new Configuration object for each model. These
objects are pretty beefy, and they increase the size of each pickle by a
full kilobyte (and of course they increase memory usage even when pickle is
not involved). AFAICS, these objects are only used when assigning values to
file-type fields, and it's easy enough to rewrite the logic so that it still
works when the model's `_configuration` field is None.