Dockerfile.ui: improve cacheability
Since Docker doesn't know which programs use which environment variables,
any `RUN` steps that follow an `ARG` step will not be loaded from the cache
if the value of the `ARG` changes. In this case, the `yarn` command will
not be loaded from the cache if any of the `ARG` values change, even though
(to my knowledge) it does not use any of those values.

Fix tihs by moving the `ARG` steps as far down as possible.